+++ src/constants/addresses.ts
@@ -138,10 +138,10 @@ export const UNISWAP_UTILS_ADDRESSES: AddressMap = {
 }
 export const KROMATIKA_ROUTER_ADDRESSES: AddressMap = {
   [SupportedChainId.ARBITRUM_ONE]: '0x79ba1CFF3998D7ce3DF452c3Fd6FCf817971Ea39',
-  [SupportedChainId.POLYGON]: '0x914A50A598700Be1E10B671f8C66a94cFe5c928B',
+  [SupportedChainId.POLYGON]: '0x4CE76357348B65e6D1002c212B9DC415Ac7F872d',
 }
 export const KROMATIKA_METASWAP_ADDRESSES: AddressMap = {
-  [SupportedChainId.POLYGON]: '0x67b1844c457f1770fE9588da116F2E1149D73B98',
+  [SupportedChainId.POLYGON]: '0xE10F5F77CF90c99976BceE524Fbf88504A2e6616',
 }
 export const BICONOMY_DAPP_API: AddressMap = {
   [SupportedChainId.POLYGON]: 'lD1x8FLPD.45318b65-8ab0-45c7-b59c-2f73137fb751',
