+++ src/hooks/useBestV3Trade.test.ts
@@ -1,10 +1,10 @@
 import { renderHook } from '@testing-library/react-hooks'
 import { CurrencyAmount, TradeType } from '@uniswap/sdk-core'
 import { DAI, USDC } from 'constants/tokens'
-import { V3TradeState } from 'state/routing/types'
 import { useRoutingAPIEnabled } from 'state/user/hooks'
+import { V3TradeState } from 'state/validator/types'
 
-import { useRoutingAPITrade } from '../state/routing/useRoutingAPITrade'
+import { useRoutingAPITrade } from '../state/validator/useRoutingAPITrade'
 import { useBestV3Trade } from './useBestV3Trade'
 import { useClientSideV3Trade } from './useClientSideV3Trade'
 import useDebounce from './useDebounce'
