+++ src/state/index.ts
@@ -13,11 +13,10 @@ import market from './market/reducer'
 import mint from './mint/reducer'
 import mintV3 from './mint/v3/reducer'
 import multicall from './multicall/reducer'
-import { routingApi0x } from './quote/slice'
-import { routingApiInch } from './routing/slice'
 import swap from './swap/reducer'
 import transactions from './transactions/reducer'
 import user from './user/reducer'
+import { routingApiKromatika } from './validator/slice'
 
 const PERSISTED_KEYS: string[] = ['user', 'transactions', 'lists']
 
@@ -36,14 +35,12 @@ const store = configureStore({
     lists,
     logs,
     [dataApi.reducerPath]: dataApi.reducer,
-    [routingApi0x.reducerPath]: routingApi0x.reducer,
-    [routingApiInch.reducerPath]: routingApiInch.reducer,
+    [routingApiKromatika.reducerPath]: routingApiKromatika.reducer,
   },
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({ thunk: true })
       .concat(dataApi.middleware)
-      .concat(routingApi0x.middleware)
-      .concat(routingApiInch.middleware)
+      .concat(routingApiKromatika.middleware)
       .concat(save({ states: PERSISTED_KEYS, debounce: 1000 })),
   preloadedState: load({ states: PERSISTED_KEYS, disableWarnings: process.env.NODE_ENV === 'test' }),
 })
