+++ src/state/swap/hooks.tsx
@@ -31,8 +31,8 @@ import { ReactNode, useCallback, useEffect, useMemo, useState } from 'react'
 import { useAppDispatch, useAppSelector } from 'state/hooks'
 import { tryParseTick } from 'state/mint/v3/utils'
 import { useSingleCallResult } from 'state/multicall/hooks'
-import { V3TradeState } from 'state/routing/types'
 import { useNetworkGasPrice, useUserTickOffset, useUserTickSize } from 'state/user/hooks'
+import { V3TradeState } from 'state/validator/types'
 
 import { V2_FACTORY_ADDRESSES } from '../../constants/addresses'
 import { useCurrency } from '../../hooks/Tokens'
